New: attribute 'multiselection' added to menu items
The presense of the attribute denotes the xul element support tab
multiselection and can be set true/false if there are some tabs
multiselected. This will be used to highlight the xul elements (i.e. menu
items).
This attribute wont be forward compatible once multiselection lands in
Firefox (as 'multiselected' attribute probably will be).